Using DEFINEs
I want to be able to switch between a dummy particle counter and the
real thing. I also want to change the sleep time when I do this.

I can change the process with a runtime IF:

IF
  TEST.RUN
    dummy.dylos (...)
  TRUE
    wait.for.dylos (...)

However, I can't do that to conditionally define a top-level constant.

So, I use

which suffices.