Adding in the config.
The boards I used on this project had Optiboot 4 (I think... some new
version number). This meant that I had to use 115200 for the baud rate,
as opposed to 57600 (like on the first board).

Adding the conf files to mark that change, and also to note where the
code should go into memory. This way, if I have to force things later, I
can.